move logic from normalize to linked items decorator
